feat(BaseService): return non-JSON responses via 'result' and DetailedResponse.Result
BREAKING CHANGE: This change to the BaseService.Request method introduces
an incompatibility with respect to the 'result' parameter.
Projects generated with the SDK generator v3.5.0 and below should
continue using version 3.x of the Go core.
Any code generated with the SDK generator version 3.6.0 or above, should use
this new version 4.0.0 of the Go core.

Note: this commit contains only a trivial change and the
breaking change message actually applies to the previous commit
with the same commit message.