feat: add support for gzip compression of request bodies
Fixes arf/planning-sdk-squad#2185

This commit adds support for peforming gzip compression
of request bodies, and consists of the following:
1. Two new functions: NewGzipCompressionReader() and
NewGzipDecompressionReader().  Each function will wrap
an existing io.Reader to provide compression/decompression
using a filter pattern.

2. A new field "EnableGzipCompression" was added to the
RequestBuilder struct to indicate whether or not request
bodies should be gzip-compressed.

3. RequestBuilder.Build() was updated to honor the new flag
and (if enabled) wrap the Body field in a gzip compression filter
prior to constructing the http.Request instance.

4. A new field "EnableGzipCompression" was added to the
ServiceOptions struct, as well as new function to support this
field as an externally configuratable option (<svcname>_ENABLE_GZIP)