feat: support use of Context with RequestBuilder
Fixes arf/planning-sdk-squad#2230
Fixes #77

This commit adds support for setting a Context
instance on the RequestBuilder struct.  This Context instance
will then be associated with the http.Request instance constructed
by the RequestBuilder.Build() method.
The Context can be used to define a timeout/deadline for a request
or to cancel an in-flight request.