feat(logger): support logging to stdout and stderr
Currently, all logs performed by the default logger in the Go core will print to Stdout.
However, error messages should be logged to Stderr. This PR implements a change to the logger
that sends error logs to Stderr and all other logs to Stdout.

Additionally, this PR adds a log statement to the request builder to log the full request
for debugging purposes.