chore: fix govet copylocks error
This error occures when there is a mutex in the structure and using a method with a value reviever. In this case the struct will be copied on the method call and it will create a new mutex therefore our previous lock became useless.

From the go vet doc: "copylocks: check for locks erroneously passed by value"