feat: add FileWithMetadata type to the core
FileWithMetadata is a struct that contains the binary data within a file, along
with specific metadata: the filename and the content type. This is a type that
the generated code occasionally uses. Currently, we generate this type as a
model on a specific service but because it is service-independent, we want to
move it to the core. This adds it as a supported type within the core. Although
the generator won't be able to change this type in options models until the next
major relase, it will use this type in the generated unmarshall functions, which
are currently broken. This will be fixed in the generator after this feature is
released.

The new "unmarshaler" in the core includes code to parse the JSON
string and read the data from the file while creating the struct. Generated
unmarshaller on service structs will invoke this new, working unmarshaller.