fix: use consistent retry behavior for 5xx status codes
This commit modifies the Go core's IBMCloudSDKRetryPolicy()
function so that retries associated with 5xx status codes
will behave the same as (for example) a 429 status code.
Specifically, this means that when retries are exhausted
as the result of a 5xx status code, the Go core will return
a non-nil DetailedResponse object, similar to a 429 status
code and also similar to a "retries disabled" scenario.
Previously, a request such as this would result in a non-nil
error return value, but a nil DetailedResponse so it was
impossible for the caller to inspect the response status code
and response headers.