fix: BaseService.Request invoked without result does not close http.Response (#176)
Fixes: https://github.com/IBM/go-sdk-core/issues/175

This commit modifies BaseService.Request() so that even if a response is not expected (i.e. the `result` parameter is nil), we'll close the response body to guard against leaking streams.   When the response body is not closed, the connection is not returned to "idle" state, which then results in excessive memory use and a quick MaxConnsPerHost limit exhaustion.

Signed-off-by: Rafał Bigaj <rafal.bigaj@pl.ibm.com>

* fix: additional tweaks

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>