fix(errors): support error checking methods for new problem structures
Adds previously returned errors to the "Unwrap" chain for checking against
in methods like "errors.Is()" and adds an implementation of "Is" to Problem
types to enable equality checking against problems.

Also, adds all of the previously returned errors back into the problem
structures to maintain full compatibility with downstream code.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>