fix(errors): restore matching against all original errors (#213)
I missed an error origination point that needed to be converted to
the new format. Also, this is maybe overly cautious but I added
back in the original error instances as native caused by errors
for all instances. Before, I intentionally left alone places that
held an error origination from `fmt.Errorf()` because I don't know
that users can really match against those, but I figured we might as
well be as compatibility-friendly as we can be, since it doesn't cost
us anything.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>