fix(BaseService): allow User-Agent to be set via service-level headers
This commit modifies the Go core's BaseService.Request() method
slightly so that when setting service-level custom headers on
the outbound request, it will now replace any existing header
with the service-level header value, rather than add the
service-level header value to the existing value that might
already exist.
The net effect of this is that a User-Agent header set
as a service-level default header will now override
a User-Agent header set by an SDK project through its
"common.GetSdkHeaders()" function.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>